Refactoring and Reformat all project files

+++ app/src/main/java/net/iGap/Config.java
@@ -29,8 +29,8 @@ public class Config {
     public static final int LIMIT_GET_HISTORY_NORMAL = 50;
     public static final int LIMIT_LOAD_ROOM = 50;
     public static final int STORE_MESSAGE_POSITION_LIMIT = 2; // count of message from end of list
-    public static final int FAST_START_PAGE_TIME = (int) 50;
-    public static final int LOW_START_PAGE_TIME = (int) 200;
+    public static final int FAST_START_PAGE_TIME = 50;
+    public static final int LOW_START_PAGE_TIME = 200;
     public static final int PHONE_CONTACT_MAX_COUNT_LIMIT = 5000;
     public static final long drIgapPeerId = 2297310;
